+++ api/test/e2e/impact/mocks/actual-vs-scenario-preconditions/new-material-intervention.preconditions.ts
@@ -8,8 +8,8 @@ import {
   ScenarioIntervention,
 } from 'modules/scenario-interventions/scenario-intervention.entity';
 import {
-  SourcingLocation,
   SOURCING_LOCATION_TYPE_BY_INTERVENTION,
+  SourcingLocation,
 } from 'modules/sourcing-locations/sourcing-location.entity';
 import { Supplier } from 'modules/suppliers/supplier.entity';
 import { Unit } from 'modules/units/unit.entity';
@@ -24,25 +24,24 @@ import {
   createSourcingRecord,
   createSupplier,
   createUnit,
-} from '../../../entity-mocks';
-import { INDICATOR_TYPES } from '../../../../src/modules/indicators/indicator.entity';
+} from '../../../../entity-mocks';
+import { INDICATOR_TYPES } from '../../../../../src/modules/indicators/indicator.entity';
 import { Scenario } from 'modules/scenarios/scenario.entity';
 
-export async function createNewSupplierInterventionPreconditions(
+export async function createNewMaterialInterventionPreconditions(
   customScenario?: Scenario,
   customIndicator?: Indicator,
   customMaterials?: Record<string, Material>,
 ): Promise<{
   indicator: Indicator;
   scenarioIntervention: ScenarioIntervention;
+  replacingMaterials: Record<string, Material>;
+  replacedMaterials: Record<string, Material>;
 }> {
   const adminRegion: AdminRegion = await createAdminRegion({
     name: 'India',
   });
-  const unit: Unit = await createUnit({
-    name: 'defFakeUnit',
-    shortName: 'fakeUnitDef',
-  });
+  const unit: Unit = await createUnit({ shortName: 'fakeUnit' });
   const indicator: Indicator = customIndicator
     ? customIndicator
     : await createIndicator({
@@ -51,36 +50,32 @@ export async function createNewSupplierInterventionPreconditions(
         nameCode: INDICATOR_TYPES.DEFORESTATION,
       });
 
-  const textile: Material = customMaterials
-    ? customMaterials.textile
-    : await createMaterial({ name: 'Textile' });
+  const textile: Material = await createMaterial({ name: 'Textile' });
 
-  const wool: Material = customMaterials
-    ? customMaterials.wool
-    : await createMaterial({
-        name: 'Wool',
-        parent: textile,
-      });
-  const cotton: Material = customMaterials
-    ? customMaterials.cotton
-    : await createMaterial({
-        name: 'Cotton',
-        parent: textile,
-      });
+  const wool: Material = await createMaterial({
+    name: 'Wool',
+    parent: textile,
+  });
+  const cotton: Material = await createMaterial({
+    name: 'Cotton',
+    parent: textile,
+  });
+
+  const linen: Material = await createMaterial({
+    name: 'Linen',
+    parent: textile,
+  });
 
   const businessUnit: BusinessUnit = await createBusinessUnit({
     name: 'Fake Business Unit',
   });
 
-  const supplierA: Supplier = await createSupplier({
-    name: 'Supplier A',
-  });
-
-  const supplierB: Supplier = await createSupplier({
-    name: 'Supplier B',
+  const supplier: Supplier = await createSupplier({
+    name: 'Fake Supplier',
   });
 
   // Scenario pre-conditions
+
   const scenarioIntervention: ScenarioIntervention = customScenario
     ? await createScenarioIntervention({ scenario: customScenario })
     : await createScenarioIntervention();
@@ -102,7 +97,7 @@ export async function createNewSupplierInterventionPreconditions(
     {
       material: cotton,
       businessUnit,
-      t1Supplier: supplierA,
+      t1Supplier: supplier,
       adminRegion,
     },
   );
@@ -110,7 +105,7 @@ export async function createNewSupplierInterventionPreconditions(
   const woolSourcingLocation: SourcingLocation = await createSourcingLocation({
     material: wool,
     businessUnit,
-    t1Supplier: supplierA,
+    t1Supplier: supplier,
     adminRegion,
   });
 
@@ -120,7 +115,7 @@ export async function createNewSupplierInterventionPreconditions(
     await createSourcingLocation({
       material: cotton,
       businessUnit,
-      t1Supplier: supplierA,
+      t1Supplier: supplier,
       adminRegion,
       scenarioInterventionId: scenarioIntervention.id,
       interventionType: SOURCING_LOCATION_TYPE_BY_INTERVENTION.CANCELED,
@@ -130,7 +125,7 @@ export async function createNewSupplierInterventionPreconditions(
     await createSourcingLocation({
       material: wool,
       businessUnit,
-      t1Supplier: supplierA,
+      t1Supplier: supplier,
       adminRegion,
       scenarioInterventionId: scenarioIntervention.id,
       interventionType: SOURCING_LOCATION_TYPE_BY_INTERVENTION.CANCELED,
@@ -140,39 +135,29 @@ export async function createNewSupplierInterventionPreconditions(
     await createSourcingLocation({
       material: wool,
       businessUnit,
-      t1Supplier: supplierA,
+      t1Supplier: supplier,
       adminRegion,
       scenarioInterventionId: scenarioInterventionInactive.id,
       interventionType: SOURCING_LOCATION_TYPE_BY_INTERVENTION.CANCELED,
     });
 
-  // Sourcing location for the Replacing Supplier
+  // Sourcing location for the Replacing Material
 
-  const woolSourcingLocationReplacing: SourcingLocation =
+  const linenSourcingLocationReplacing: SourcingLocation =
     await createSourcingLocation({
-      material: wool,
+      material: linen,
       businessUnit,
-      t1Supplier: supplierB,
+      t1Supplier: supplier,
       adminRegion,
       scenarioInterventionId: scenarioIntervention.id,
       interventionType: SOURCING_LOCATION_TYPE_BY_INTERVENTION.REPLACING,
     });
 
-  const cottonSourcingLocationReplacing: SourcingLocation =
+  const linenSourcingLocationReplacingInactive: SourcingLocation =
     await createSourcingLocation({
-      material: cotton,
+      material: linen,
       businessUnit,
-      t1Supplier: supplierB,
-      adminRegion,
-      scenarioInterventionId: scenarioIntervention.id,
-      interventionType: SOURCING_LOCATION_TYPE_BY_INTERVENTION.REPLACING,
-    });
-
-  const cottonSourcingLocationReplacingInactive: SourcingLocation =
-    await createSourcingLocation({
-      material: cotton,
-      businessUnit,
-      t1Supplier: supplierB,
+      t1Supplier: supplier,
       adminRegion,
       scenarioInterventionId: scenarioInterventionInactive.id,
       interventionType: SOURCING_LOCATION_TYPE_BY_INTERVENTION.REPLACING,
@@ -187,43 +172,37 @@ export async function createNewSupplierInterventionPreconditions(
 
   const indicatorRecordWool: IndicatorRecord = await createIndicatorRecord({
     indicator,
-    value: 1200,
+    value: 1500,
   });
 
   const indicatorRecordCottonCancelled: IndicatorRecord =
     await createIndicatorRecord({
       indicator,
-      value: -1200,
+      value: -800,
     });
 
   const indicatorRecordWoolCancelled: IndicatorRecord =
     await createIndicatorRecord({
       indicator,
-      value: -1200,
+      value: -1000,
+    });
+
+  const indicatorRecordLinenReplacing: IndicatorRecord =
+    await createIndicatorRecord({
+      indicator,
+      value: 750,
     });
 
   const indicatorRecordWoolCancelledInactive: IndicatorRecord =
     await createIndicatorRecord({
       indicator,
-      value: -123,
+      value: -100,
     });
 
-  const indicatorRecordWoolReplacing: IndicatorRecord =
+  const indicatorRecordLinenReplacingInactive: IndicatorRecord =
     await createIndicatorRecord({
       indicator,
-      value: 1000,
-    });
-
-  const indicatorRecordCottonReplacing: IndicatorRecord =
-    await createIndicatorRecord({
-      indicator,
-      value: 900,
-    });
-
-  const indicatorRecordCottonReplacingInactive: IndicatorRecord =
-    await createIndicatorRecord({
-      indicator,
-      value: 567,
+      value: 50,
     });
 
   // Sourcing Records + Indicator Records for Real Sourcing Locations
@@ -264,21 +243,20 @@ export async function createNewSupplierInterventionPreconditions(
 
   await createSourcingRecord({
     year: 2020,
-    indicatorRecords: [indicatorRecordWoolReplacing],
-    sourcingLocation: woolSourcingLocationReplacing,
+    indicatorRecords: [indicatorRecordLinenReplacing],
+    sourcingLocation: linenSourcingLocationReplacing,
   });
 
   await createSourcingRecord({
     year: 2020,
-    indicatorRecords: [indicatorRecordCottonReplacing],
-    sourcingLocation: cottonSourcingLocationReplacing,
+    indicatorRecords: [indicatorRecordLinenReplacingInactive],
+    sourcingLocation: linenSourcingLocationReplacingInactive,
   });
 
-  await createSourcingRecord({
-    year: 2020,
-    indicatorRecords: [indicatorRecordCottonReplacingInactive],
-    sourcingLocation: cottonSourcingLocationReplacingInactive,
-  });
-
-  return { indicator, scenarioIntervention };
+  return {
+    indicator,
+    scenarioIntervention,
+    replacedMaterials: { wool, cotton },
+    replacingMaterials: { linen },
+  };
 }
