+++ api/test/e2e/impact/mocks/actual-vs-scenario-responses/mixed-interventions-scenario.response.ts
