+++ api/test/e2e/impact/mocks/actual-vs-scenario-responses/new-coefficients-intervention.response.ts
