+++ api/test/e2e/impact/mocks/scenario-vs-scenario-preconditions/mixed-scenarios.preconditions.ts
@@ -26,8 +26,8 @@ import {
   createSourcingRecord,
   createSupplier,
   createUnit,
-} from '../../../entity-mocks';
-import { INDICATOR_TYPES } from '../../../../src/modules/indicators/indicator.entity';
+} from '../../../../entity-mocks';
+import { INDICATOR_TYPES } from '../../../../../src/modules/indicators/indicator.entity';
 
 export async function createMixedScenariosPreconditions(): Promise<{
   indicator: Indicator;
